Fix a race condition in the creation of ShowmapSync jobs.
Before this fix, the creator would only see if raw_round_polls
existed for the previous round for the given challenge set;
however, this doesn't account for the possibility that traffic
may be processed late (for example, traffic from round 2 may be
processed on round 4). When this occurs no ShowmapSync job would
be created for those RawRoundPolls because none would actually
exist yet.